Remove hapi.py from RADIS files
This requires upgrading the version of HAPI that RADIS is using. Before
`hapi.py` was at version 1.1.0.6 but this version is not available on
PyPI. As well, the earliest versions on PyPI do not support Python 3.6,
so it would be good to go ahead and use the latest HAPI version if
possible.

Some tests are failing due to the upgraded HAPI version it seems like.

Refs: #151